Bug Backend Product Model
Change long type to float of attribute weight.

+++ backend/src/main/java/com/example/weight_inspection/models/Product.java
@@ -16,7 +16,7 @@ public class Product {
     @NotNull
     private String refference;
     @NotNull
-    private Long weigth;
+    private Float weigth;
 
     public Product() {}
